update model and controller
