+++ src/resources/stylesheets/components/_toast.scss
@@ -15,10 +15,10 @@
   opacity: 0.9;
 }
 
-.toast.success {
+.toasts.success {
   background-color: #4caf50;
 }
 
-.toast.error {
+.toasts.error {
   background-color: #f44336;
 }
